+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to thsi project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.4.2] - 2018-05-02
+### Added
+- Added "help" with search terms
+- Added "repeat" command to repeat a task
+- Added description of todo.txt in "help todo.txt"
+
 ### Changed
 - 'addx' command supports '-p' prompt